Move templating helpers to own file
